[tradeoff-analytics] Fix models by adding statusCause to the Solution POJO
