+++ src/main/java/com/ibm/watson/developer_cloud/tradeoff_analytics/v1/model/Problem.java
@@ -17,20 +17,16 @@ package com.ibm.watson.developer_cloud.tradeoff_analytics.v1.model;
 import java.util.List;
 
 import com.ibm.watson.developer_cloud.service.model.GenericModel;
+import com.ibm.watson.developer_cloud.tradeoff_analytics.v1.TradeoffAnalytics;
 import com.ibm.watson.developer_cloud.tradeoff_analytics.v1.model.column.Column;
 
 /**
- * The Class Problem.
+ * The Problem representation to be used by the {@link TradeoffAnalytics} V1 service
  */
 public class Problem extends GenericModel {
 
-  /** The columns. */
   private List<Column> columns;
-
-  /** The options. */
   private List<Option> options;
-
-  /** The subject. */
   private String subject;
 
   /**
