+++ src/main/java/com/ibm/watson/developer_cloud/tradeoff_analytics/v1/model/Solution.java
@@ -18,13 +18,67 @@ import java.util.List;
 
 import com.google.gson.annotations.SerializedName;
 import com.ibm.watson.developer_cloud.service.model.GenericModel;
-
-
+import com.ibm.watson.developer_cloud.tradeoff_analytics.v1.TradeoffAnalytics;
 
 /**
- * The Class Solution.
+ * The Solution used by the {@link TradeoffAnalytics} v1 service
  */
 public class Solution extends GenericModel {
+
+  /**
+   * Represent the error when the {@link Solution#status} is INCOMPLETE.
+   */
+  public class StatusCause {
+
+    @SerializedName("error_code")
+    private String errorCode;
+
+    private String message;
+    private List<String> tokens;
+
+    /**
+     * @return the errorCode
+     */
+    public String getErrorCode() {
+      return errorCode;
+    }
+
+    /**
+     * @return the message
+     */
+    public String getMessage() {
+      return message;
+    }
+
+    /**
+     * @return the tokens
+     */
+    public List<String> getTokens() {
+      return tokens;
+    }
+
+    /**
+     * @param errorCode the errorCode to set
+     */
+    public void setErrorCode(String errorCode) {
+      this.errorCode = errorCode;
+    }
+
+    /**
+     * @param message the message to set
+     */
+    public void setMessage(String message) {
+      this.message = message;
+    }
+
+    /**
+     * @param tokens the tokens to set
+     */
+    public void setTokens(List<String> tokens) {
+      this.tokens = tokens;
+    }
+
+  }
 
   @SerializedName("shadow_me")
   private List<String> shadowMe;
@@ -34,8 +88,11 @@ public class Solution extends GenericModel {
   @SerializedName("solution_ref")
   private String solutionRef;
 
+
   private String status;
 
+  @SerializedName("status_cause")
+  private StatusCause statusCause;
 
   /**
    * Gets the solution ids of those that shadow this solution.
@@ -64,6 +121,7 @@ public class Solution extends GenericModel {
     return solutionRef;
   }
 
+
   /**
    * Gets the status.
    * 
@@ -71,6 +129,13 @@ public class Solution extends GenericModel {
    */
   public String getStatus() {
     return status;
+  }
+
+  /**
+   * @return the statusCause
+   */
+  public StatusCause getStatusCause() {
+    return statusCause;
   }
 
   /**
@@ -109,26 +174,10 @@ public class Solution extends GenericModel {
     this.status = status;
   }
 
-
   /**
-   * With shadowME.
-   * 
-   * @param shadowMe the highlights
-   * @return the solution
+   * @param statusCause the statusCause to set
    */
-  public Solution withShadowMe(List<String> shadowMe) {
-    this.shadowMe = shadowMe;
-    return this;
-  }
-
-  /**
-   * With shadows.
-   * 
-   * @param shadows the shadows
-   * @return the solution
-   */
-  public Solution withShadows(List<String> shadows) {
-    this.shadows = shadows;
-    return this;
+  public void setStatusCause(StatusCause statusCause) {
+    this.statusCause = statusCause;
   }
 }
