+++ src/main/java/com/ibm/watson/developer_cloud/util/ResponseUtil.java
@@ -112,7 +112,6 @@ public class ResponseUtil {
    */
   public static <T extends GenericModel> T getObject(Response response, Class<T> type) {
     final String jsonString = getString(response);
-    System.out.println(jsonString);
     final T model = GsonSingleton.getGson().fromJson(jsonString, type);
     return model;
   }
